requirements.txt
Django==1.11
Pillow==5.0.0
boto==2.48.0
boto3==1.6.19
botocore==1.9.19
coverage==4.5.1
dj-database-url==0.5.0
django-filter==1.1.0
django-forms-bootstrap==3.1.0
django-storages==1.6.6
docutils==0.14
gunicorn==19.7.1
jmespath==0.9.3
psycopg2==2.7.4
python-dateutil==2.6.1
python-http-client==3.0.0
pytz==2018.3
s3transfer==0.1.13
sendgrid==5.3.0
stripe==1.79.1
whitenoise==3.3.1
